move precise iv curve fixtures into test_singlediode
