+++ pvlib/tests/test_singlediode.py
@@ -3,10 +3,12 @@ testing single-diode methods using JW Bishop 1988
 """
 
 import numpy as np
+import pandas as pd
 from pvlib import pvsystem
 from pvlib.singlediode import (bishop88_mpp, estimate_voc, VOLTAGE_BUILTIN,
                                bishop88, bishop88_i_from_v, bishop88_v_from_i)
 import pytest
+from .conftest import DATA_DIR
 
 POA = 888
 TCELL = 55
@@ -69,18 +71,88 @@ def test_newton_fs_495(method, cec_module_fs_495):
     assert np.isclose(pvs_ixx, ixx)
 
 
+def build_precise_iv_curve_dataframe(file_csv, file_json):
+    """
+    Reads a precise IV curve parameter set CSV and JSON to create a data frame
+    with these columns: ``Index``, ``photocurrent``, ``saturation_current``,
+    ``resistance_series``, ``resistance_shunt``, ``n``, ``cells_in_series``,
+    ``Voltages``, ``Currents``, ``v_oc``, ``i_sc``, ``v_mp``, ``i_mp``,
+    ``p_mp``, ``i_x``, ``i_xx`, ``Temperature``, ``Irradiance``,
+    ``Sweep Direction``, ``Datetime``, ``Boltzman``, ``Electron Charge``, and
+    ``Vth``. The columns ``Irradiance``, ``Sweep Direction`` are null or empty
+    strings.
+
+    Parameters
+    ----------
+    file_csv: str
+        Path to a CSV file of iv curve parameter sets.
+
+    file_json: str
+        Path to a JSON file of precise iv curves.
+
+    Returns
+    -------
+        A data frame.
+    """
+    params = pd.read_csv(file_csv)
+    curves = pd.read_json(file_json)
+    curves = pd.DataFrame(curves['IV Curves'].values.tolist())
+    joined = params.merge(curves, on='Index', how='inner',
+                          suffixes=(None, '_drop'), validate='one_to_one')
+    joined = joined[(c for c in joined.columns if not c.endswith('_drop'))]
+
+    joined['Boltzman'] = 1.380649e-23
+    joined['Electron Charge'] = 1.60217663e-19
+    joined['Vth'] = (
+        joined['Boltzman'] * joined['Temperature']
+        / joined['Electron Charge']
+    )
+
+    return joined
+
+
+@pytest.fixture(scope='function', params=[
+    {
+        'csv': f'{DATA_DIR}/precise_iv_curves_parameter_sets1.csv',
+        'json': f'{DATA_DIR}/precise_iv_curves1.json'
+    },
+    {
+        'csv': f'{DATA_DIR}/precise_iv_curves_parameter_sets2.csv',
+        'json': f'{DATA_DIR}/precise_iv_curves2.json'
+    }
+], ids=[1, 2])
+def precise_iv_curves(request):
+    file_csv, file_json = request.param['csv'], request.param['json']
+    return build_precise_iv_curve_dataframe(file_csv, file_json)
+
+
+@pytest.fixture(scope='function', params=[
+    {
+        'csv': f'{DATA_DIR}/precise_iv_curves_parameter_sets1.csv',
+        'json': f'{DATA_DIR}/precise_iv_curves1_log_spacing.json'
+    },
+    {
+        'csv': f'{DATA_DIR}/precise_iv_curves_parameter_sets2.csv',
+        'json': f'{DATA_DIR}/precise_iv_curves2_log_spacing.json'
+    }
+], ids=[1, 2])
+def precise_iv_curves_log_spacing(request):
+    file_csv, file_json = request.param['csv'], request.param['json']
+    return build_precise_iv_curve_dataframe(file_csv, file_json)
+
+
 @pytest.mark.parametrize('method', ['lambertw', 'brentq', 'newton'])
 def test_singlediode_precision(method, precise_iv_curves):
     """
     Tests the accuracy of singlediode. ivcurve_pnts is not tested.
     """
     pc = precise_iv_curves
-    il, io, rs, rsh, nnsvt = (
+    x = (
         pc['photocurrent'], pc['saturation_current'],
         pc['resistance_series'], pc['resistance_shunt'],
         pc['n'] * pc['cells_in_series'] * pc['Vth']
     )
-    outs = pvsystem.singlediode(il, io, rs, rsh, nnsvt, method=method)
+    outs = pvsystem.singlediode(*x, method=method)
 
     assert np.allclose(pc['i_sc'], outs['i_sc'], atol=1e-10, rtol=0)
     assert np.allclose(pc['v_oc'], outs['v_oc'], atol=1e-10, rtol=0)
